Add basic caching to speed up adding new bookmarks
Since have to (well should do, to be nice) rate limit quite a lot it can
be slow to add new bookmarks to Pinboard when we are checking Pinboard
for every single StackExchange fav. Keep a local cache of items already
added to Pinboard.

Some things I have to do (note to self):

- Check for success on Pinboard add before adding to cache
- Consider removing bookmarks on Pinboard for unfaved StackExchange
  items?